src/main/java/com/project/recipeproject/controllers/RecipeController.java
package com.project.recipeproject.controllers;

import com.project.recipeproject.commands.RecipeCommand;
import com.project.recipeproject.services.RecipeService;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.*;

@Slf4j
@Controller
@RequestMapping("/recipe")
public class RecipeController {
    private final RecipeService recipeService;

    public RecipeController(RecipeService recipeService) {
        this.recipeService = recipeService;
    }

    @GetMapping("/{id}/show")
    public String showById(@PathVariable String id, Model model) {
        model.addAttribute("recipe", recipeService.findById(Long.parseLong(id)));
        return "recipe/show";
    }

    @GetMapping("/new")
    public String createRecipe(Model model) {
        model.addAttribute("recipe", new RecipeCommand());
        return "recipe/edit";
    }

    @GetMapping("/{id}/update")
    public String updateRecipe(@PathVariable String id, Model model) {
        model.addAttribute("recipe", recipeService.findCommandById(Long.parseLong(id)));
        return "recipe/edit";
    }

    @PostMapping("/change")
    public String saveOrUpdate(@ModelAttribute RecipeCommand recipeCommand) {
        var savedCommand = recipeService.saveRecipe(recipeCommand);
        return "redirect:/recipe/" + savedCommand.getId() + "/show";
    }

    @GetMapping("/{id}/delete")
    public String delete(@PathVariable String id) {
        recipeService.deleteRecipe(Long.parseLong(id));
        return "redirect:/";
    }
}
